+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Now if any exception occurs, it can be sent with stack trace as a file to
+ a chat.
+
 ### Changed
  - Now tglogger checks `DJANGO_SETTINGS_MODULE` environment variable instead of
  the presence of `django`.